Finished Course Block Info!
I added functionality for classes that have different locations and times on different days, and formatted the week schedule display into a simple table so it is easier to read as the window is resized.

The only bug is the background and border colors working for only some colors but not all (the only one that completely works is lake herrick). I don't know the cause since I'm still very new to tailwind syntax.