Dynamic Navbar
Updated the Navbar component so that it can take in a keyword as a prop to conditionally render elements based on the needs of the page. If the prop is left off, it will default to the normal Navbar. Specifically added a variant for the schedules page using keyword = "schedules". Updated schedules/layout.tsx to use the dynamic Navbar.