Fixed spring boot from crashing
I solved the issue with it crashing we had in the meeting by saving the course before the course section. To make debugging a little easier we now have a toString method for websracbing's course objects which revealed that sec sometimes returns '' which crashes the program. So as a hot fix I just set it to 99. I havent tried to load all the values as it looks like its going to take a long time but if you want to watch the program read the data you just have to run it.